Added root accounts

+++ src/app/Http/Controllers/Api/v2/AccountApiController.php
@@ -201,9 +201,19 @@ class AccountApiController extends Controller
     public function delete(Request $request, int $accountId)
     {
         $account = $this->getAuthorizedAccount($request, $accountId);
+        if ($account->isAdministrator()) {
+            abort(400, 'You cannot delete an administrator\'s account.');
+        }
+
         $this->_accountManager->delete($account);
 
         $redirectUrl = route('logout');
+        if ($request->user()->id !== $account->id &&
+            $request->user()->isAdministrator()) {
+            // This is only possible if an administrator deleted the account.
+            $redirectUrl = route('account.index');
+        }
+
         if (! $request->ajax()) {
             return redirect($redirectUrl);
         }
